+++ loguru/_logger.py
@@ -1,4 +1,5 @@
-"""
+"""Core logging functionalities of the `Loguru` library.
+
 .. References and links rendered by Sphinx are kept here as "module documentation" so that they can
    be used in the ``Logger`` docstrings but do not pollute ``help(logger)`` output.
 
@@ -89,6 +90,7 @@ import functools
 import logging
 import re
 import sys
+import threading
 import warnings
 from collections import namedtuple
 from inspect import isclass, iscoroutinefunction, isgeneratorfunction
@@ -116,7 +118,7 @@ else:
     from pathlib import PurePath as PathLike
 
 
-Level = namedtuple("Level", ["name", "no", "color", "icon"])
+Level = namedtuple("Level", ["name", "no", "color", "icon"])  # noqa: PYI024
 
 start_time = aware_now()
 
@@ -193,15 +195,18 @@ class Core:
         self.activation_list = []
         self.activation_none = True
 
+        self.thread_locals = threading.local()
         self.lock = create_logger_lock()
 
     def __getstate__(self):
         state = self.__dict__.copy()
+        state["thread_locals"] = None
         state["lock"] = None
         return state
 
     def __setstate__(self, state):
         self.__dict__.update(state)
+        self.thread_locals = threading.local()
         self.lock = create_logger_lock()
 
 
@@ -661,9 +666,10 @@ class Logger:
 
         Tags which are not recognized will raise an exception during parsing, to inform you about
         possible misuse. If you wish to display a markup tag literally, you can escape it by
-        prepending a ``\`` like for example ``\<blue>``. If, for some reason, you need to escape a
-        string programmatically, note that the regex used internally to parse markup tags is
-        ``r"\\?</?((?:[fb]g\s)?[^<>\s]*)>"``.
+        prepending a ``\`` like for example ``\<blue>``. To prevent the escaping to occur, you can
+        simply double the ``\`` (e.g. ``\\<blue>`` will print a literal ``\`` before colored text).
+        If, for some reason, you need to escape a string programmatically, note that the regex used
+        internally to parse markup tags is ``r"(\\*)(</?(?:[fb]g\s)?[^<>\s]*>)"``.
 
         Note that when logging a message with ``opt(colors=True)``, color tags present in the
         formatting arguments (``args`` and ``kwargs``) are completely ignored. This is important if
@@ -1042,17 +1048,16 @@ class Logger:
             )
 
         with self._core.lock:
-            handlers = self._core.handlers.copy()
-
-            if handler_id is not None and handler_id not in handlers:
+            if handler_id is not None and handler_id not in self._core.handlers:
                 raise ValueError("There is no existing handler with id %d" % handler_id) from None
 
             if handler_id is None:
-                handler_ids = list(handlers.keys())
+                handler_ids = list(self._core.handlers)
             else:
                 handler_ids = [handler_id]
 
             for handler_id in handler_ids:
+                handlers = self._core.handlers.copy()
                 handler = handlers.pop(handler_id)
 
                 # This needs to be done first in case "stop()" raises an exception
@@ -1230,6 +1235,15 @@ class Logger:
                 if type_ is None:
                     return None
 
+                # We must prevent infinite recursion in case "logger.catch()" handles an exception
+                # that occurs while logging another exception. This can happen for example when
+                # the exception formatter calls "repr(obj)" while the "__repr__" method is broken
+                # but decorated with "logger.catch()". In such a case, we ignore the catching
+                # mechanism and just let the exception be thrown (that way, the formatter will
+                # rightly assume the object is unprintable).
+                if getattr(logger._core.thread_locals, "already_logging_exception", False):
+                    return False
+
                 if not issubclass(type_, exception):
                     return False
 
@@ -1242,8 +1256,13 @@ class Logger:
                 if from_decorator:
                     depth += 1
 
-                catch_options = [(type_, value, traceback_), depth, True] + options
-                logger._log(level, from_decorator, catch_options, message, (), {})
+                catch_options = [(type_, value, traceback_), depth, True, *options]
+
+                logger._core.thread_locals.already_logging_exception = True
+                try:
+                    logger._log(level, from_decorator, catch_options, message, (), {})
+                finally:
+                    logger._core.thread_locals.already_logging_exception = False
 
                 if onerror is not None:
                     onerror(value)
@@ -1513,10 +1532,10 @@ class Logger:
         ...     logger.patch(lambda r: r.update(record)).log(level, message)
         """
         *options, patchers, extra = self._options
-        return Logger(self._core, *options, patchers + [patcher], extra)
+        return Logger(self._core, *options, [*patchers, patcher], extra)
 
     def level(self, name, no=None, color=None, icon=None):
-        """Add, update or retrieve a logging level.
+        r"""Add, update or retrieve a logging level.
 
         Logging levels are defined by their ``name`` to which a severity ``no``, an ansi ``color``
         tag and an ``icon`` are associated and possibly modified at run-time. To |log| to a custom
@@ -1550,7 +1569,8 @@ class Logger:
         Raises
         ------
         ValueError
-            If there is no level registered with such ``name``.
+            If attempting to access a level with a ``name`` that is not registered, or if trying to
+            change the severity ``no`` of an existing level.
 
         Examples
         --------
@@ -1587,7 +1607,7 @@ class Logger:
                 )
             old_color, old_icon = "", " "
         elif no is not None:
-            raise TypeError("Level '%s' already exists, you can't update its severity no" % name)
+            raise ValueError("Level '%s' already exists, you can't update its severity no" % name)
         else:
             _, no, old_color, old_icon = self.level(name)
 
@@ -1939,10 +1959,21 @@ class Logger:
 
         (exception, depth, record, lazy, colors, raw, capture, patchers, extra) = options
 
-        frame = get_frame(depth + 2)
+        try:
+            frame = get_frame(depth + 2)
+        except ValueError:
+            f_globals = {}
+            f_lineno = 0
+            co_name = "<unknown>"
+            co_filename = "<unknown>"
+        else:
+            f_globals = frame.f_globals
+            f_lineno = frame.f_lineno
+            co_name = frame.f_code.co_name
+            co_filename = frame.f_code.co_filename
 
         try:
-            name = frame.f_globals["__name__"]
+            name = f_globals["__name__"]
         except KeyError:
             name = None
 
@@ -1968,9 +1999,7 @@ class Logger:
 
         current_datetime = aware_now()
 
-        code = frame.f_code
-        file_path = code.co_filename
-        file_name = basename(file_path)
+        file_name = basename(co_filename)
         thread = current_thread()
         process = current_process()
         elapsed = current_datetime - start_time
@@ -1990,10 +2019,10 @@ class Logger:
             "elapsed": elapsed,
             "exception": exception,
             "extra": {**core.extra, **context.get(), **extra},
-            "file": RecordFile(file_name, file_path),
-            "function": code.co_name,
+            "file": RecordFile(file_name, co_filename),
+            "function": co_name,
             "level": RecordLevel(level_name, level_no, level_icon),
-            "line": frame.f_lineno,
+            "line": f_lineno,
             "message": str(message),
             "module": splitext(file_name)[0],
             "name": name,
@@ -2067,7 +2096,7 @@ class Logger:
         __self._log("CRITICAL", False, __self._options, __message, args, kwargs)
 
     def exception(__self, __message, *args, **kwargs):  # noqa: N805
-        r"""Convenience method for logging an ``'ERROR'`` with exception information."""
+        r"""Log an ``'ERROR'```` message while also capturing the currently handled exception."""
         options = (True,) + __self._options[1:]
         __self._log("ERROR", False, options, __message, args, kwargs)
 
@@ -2076,7 +2105,9 @@ class Logger:
         __self._log(__level, False, __self._options, __message, args, kwargs)
 
     def start(self, *args, **kwargs):
-        """Deprecated function to |add| a new handler.
+        """Add a handler sending log messages to a sink adequately configured.
+
+        Deprecated function, use |add| instead.
 
         Warnings
         --------
@@ -2092,7 +2123,9 @@ class Logger:
         return self.add(*args, **kwargs)
 
     def stop(self, *args, **kwargs):
-        """Deprecated function to |remove| an existing handler.
+        """Remove a previously added handler and stop sending logs to its sink.
+
+        Deprecated function, use |remove| instead.
 
         Warnings
         --------