+++ tests/test_add_option_context.py
@@ -7,13 +7,6 @@ import pytest
 from loguru import logger
 
 
-@pytest.fixture
-def reset_start_method():
-    yield
-    multiprocessing.set_start_method(None, force=True)
-
-
-@pytest.mark.usefixtures("reset_start_method")
 def test_using_multiprocessing_directly_if_context_is_none():
     logger.add(lambda _: None, enqueue=True, context=None)
     assert multiprocessing.get_start_method(allow_none=True) is not None