+++ tests/test_exceptions_formatting.py
@@ -12,7 +12,7 @@ from loguru import logger
 
 
 def normalize(exception):
-    """Normalize exception output for reproducible test cases"""
+    """Normalize exception output for reproducible test cases."""
     if os.name == "nt":
         exception = re.sub(
             r'File[^"]+"[^"]+\.py[^"]*"', lambda m: m.group().replace("\\", "/"), exception
@@ -82,7 +82,7 @@ def normalize(exception):
 
 
 def generate(output, outpath):
-    """Generate new output file if exception formatting is updated"""
+    """Generate new output file if exception formatting is updated."""
     os.makedirs(os.path.dirname(outpath), exist_ok=True)
     with open(outpath, "w") as file:
         file.write(output)
@@ -202,6 +202,7 @@ def test_exception_ownership(filename):
     "filename",
     [
         "assertionerror_without_traceback",
+        "broken_but_decorated_repr",
         "catch_as_context_manager",
         "catch_as_decorator_with_parentheses",
         "catch_as_decorator_without_parentheses",
@@ -218,6 +219,9 @@ def test_exception_ownership(filename):
         "message_formatting_with_context_manager",
         "message_formatting_with_decorator",
         "nested_with_reraise",
+        "one_liner_recursion",
+        "recursion_error",
+        "repeated_lines",
         "syntaxerror_without_traceback",
         "sys_tracebacklimit",
         "sys_tracebacklimit_negative",
@@ -227,11 +231,14 @@ def test_exception_ownership(filename):
     ],
 )
 def test_exception_others(filename):
+    if filename == "recursion_error" and platform.python_implementation() == "PyPy":
+        pytest.skip("RecursionError is not reliable on PyPy")
+
     compare_exception("others", filename)
 
 
 @pytest.mark.parametrize(
-    "filename, minimum_python_version",
+    ("filename", "minimum_python_version"),
     [
         ("type_hints", (3, 6)),
         ("positional_only_argument", (3, 8)),
@@ -295,7 +302,7 @@ def test_invalid_format_exception_only_indented_error_message(writer, monkeypatc
 
 @pytest.mark.skipif(sys.version_info < (3, 11), reason="No builtin GroupedException")
 def test_invalid_grouped_exception_no_exceptions(writer):
-    error = MagicMock(spec=ExceptionGroup)
+    error = MagicMock(spec=ExceptionGroup)  # noqa: F821
     error.__cause__ = None
     error.__context__ = None
     error.__traceback__ = None